requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
astroid==2.4.0
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.1.7
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cryptography==2.9.2
dill==0.3.1.1
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
flask-marshmallow==0.12.0
Flask-SQLAlchemy==2.4.1
future==0.18.2
genie==20.4.1
genie.libs.clean==20.4
genie.libs.conf==20.4
genie.libs.filetransferutils==20.4
genie.libs.ops==20.4
genie.libs.parser==20.4
genie.libs.sdk==20.4
html5lib==1.0.1
idna==2.8
importlib-metadata==1.6.0
ipaddr==2.2.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpickle==1.4.1
junit-xml==1.9
lazy-object-proxy==1.4.3
lockfile==0.12.2
MarkupSafe==1.1.1
marshmallow==3.5.2
marshmallow-sqlalchemy==0.23.0
mccabe==0.6.1
msgpack==0.6.2
multidict==4.7.5
netaddr==0.7.19
netmiko==3.1.0
packaging==20.3
paramiko==2.7.1
pathspec==0.8.0
pep517==0.8.2
ply==3.11
prettytable==0.7.2
progress==1.5
psutil==5.7.0
pyasn1==0.4.8
pyats==20.4
pyats.aereport==20.4
pyats.aetest==20.4
pyats.async==20.4
pyats.connections==20.4
pyats.datastructures==20.4
pyats.easypy==20.4.1
pyats.kleenex==20.4
pyats.log==20.4.3
pyats.reporter==20.4.1
pyats.results==20.4
pyats.tcl==20.4
pyats.topology==20.4
pyats.utils==20.4
pycparser==2.20
pycryptodomex==3.9.7
pylint==2.5.0
pylint-flask==0.6
pylint-plugin-utils==0.6
PyNaCl==1.3.0
pyparsing==2.4.6
pyserial==3.4
pysmi==0.3.4
pysnmp==4.4.12
python-engineio==3.12.1
python-socketio==4.5.1
pytoml==0.1.21
PyYAML==5.3.1
requests==2.22.0
retrying==1.3.3
scp==0.13.2
six==1.14.0
SQLAlchemy==1.3.16
textfsm==1.1.0
toml==0.10.0
tqdm==4.45.0
unicon==20.4
unicon.plugins==20.4
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xmltodict==0.12.0
yamllint==1.23.0
yarl==1.4.2
zipp==3.1.0
